requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
requests==2.32.3
google-generativeai>=0.8.0
openai>=1.0.0
